aabファイル作るために変更
