+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.norikaewidget"
+        applicationId "com.aaa.norikaewidget"
         minSdk 21
         targetSdk 31
         versionCode 1
