+++ app/src/main/java/com/aaa/norikaewidget/BlankFragment.kt
@@ -1,0 +1,18 @@
+package com.aaa.norikaewidget
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.norikaewidget.R
+
+class BlankFragment : Fragment() {
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.blanklayout, container, false)
+    }
+}
