--- app/src/main/java/com/example/norikaewidget/BlankFragment.kt
@@ -1,17 +1,0 @@
-package com.example.norikaewidget
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class BlankFragment : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.blanklayout, container, false)
-    }
-}
