+++ app/src/main/res/layout/activity_time_schedule.xml
@@ -3,7 +3,7 @@
     android:layout_height="match_parent"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".TimeSchedule"
+    tools:context="com.aaa.norikaewidget.TimeSchedule"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
     <Button
